turn status_code in BaseResp from int64 to int32.

+++ cmd/api/biz/handler/utils.go
@@ -9,7 +9,7 @@ import (
 // SendResponse Package response body with status_code and status_body.
 func SendResponse(c *app.RequestContext, err error, data map[string]interface{}) {
 	Err := errno.ConvertErr(err)
-	data["status_code"] = int32(Err.ErrCode)
+	data["status_code"] = Err.ErrCode
 	data["status_msg"] = Err.ErrMsg
 	c.JSON(consts.StatusOK, data)
 }
